Rotate node log output
The console output of each log is split up in chunks of configurable
size, compressed, and when the size of the compressed output exceeds
a configurable limit, the oldest chunks are deleted.

+++ deployment/scripts/start-bootstrap.sh
@@ -1,6 +1,23 @@
 #!/bin/bash
 
+# Obtain number of lines per log file.
+log_file_lines="$1"
+[ "${log_file_lines}" -gt 0 ] || exit
+shift
+
+# Make sure that the maximal log archive size (in bytes) has been properly specified.
+max_archive_size=$1
+[ "${max_archive_size}" -gt 0 ] || exit
+shift
+
+
 cd lotus || exit
+
+# Create log directories
+bootstrap_log_dir=~/spacenet-logs/bootstrap-$(date +%Y-%m-%d-%H-%M-%S_%Z)
+faucet_log_dir=~/spacenet-logs/faucet-$(date +%Y-%m-%d-%H-%M-%S_%Z)
+mkdir -p "$bootstrap_log_dir"
+mkdir -p "$faucet_log_dir"
 
 # Kill a potentially running instance of Lotus
 tmux kill-session -t lotus
@@ -9,7 +26,7 @@ tmux new-session -d -s lotus
 # Start the Lotus daemon and import the bootstrap key.
 # Keeping the version with a custom genesis commented, in case we need to come back to it.
 #tmux send-keys "./lotus daemon --genesis=spacenet-genesis.car --profile=bootstrapper --bootstrap=false 2>&1" C-m
-tmux send-keys "./lotus daemon --profile=bootstrapper --bootstrap=false 2>&1" C-m
+tmux send-keys "./lotus daemon --profile=bootstrapper --bootstrap=false 2>&1 | ./rotate-logs.sh ${bootstrap_log_dir} ${log_file_lines} ${max_archive_size}" C-m
 mkdir -p ~/.lotus/keystore && chmod 0700 ~/.lotus/keystore
 ./lotus-shed keyinfo import spacenet-libp2p-bootstrap1.keyinfo
 echo '[Libp2p]
@@ -22,4 +39,4 @@ ListenAddresses = ["/ip4/0.0.0.0/tcp/1347"]' > ~/.lotus/config.toml
 cd ~/spacenet/faucet/ || exit
 go build -o spacenet-faucet ./cmd || exit
 tmux new-session -d -s faucet
-tmux send-keys "export LOTUS_PATH=~/.lotus && ./spacenet-faucet --web-host \"0.0.0.0:8000\" --web-allowed-origins \"*\" --web-backend-host \"https://spacenet.consensus.ninja/fund\" --filecoin-address=t1jlm55oqkdalh2l3akqfsaqmpjxgjd36pob34dqy --lotus-api-host=127.0.0.1:1234 2>&1" C-m
+tmux send-keys "export LOTUS_PATH=~/.lotus && ./spacenet-faucet --web-host \"0.0.0.0:8000\" --web-allowed-origins \"*\" --web-backend-host \"https://spacenet.consensus.ninja/fund\" --filecoin-address=t1jlm55oqkdalh2l3akqfsaqmpjxgjd36pob34dqy --lotus-api-host=127.0.0.1:1234 2>&1 | ~/lotus/rotate-logs.sh ${faucet_log_dir} ${log_file_lines} ${max_archive_size}" C-m
